added unescape function in account files
