+++ public/javascripts/escape/escape.js
@@ -1,0 +1,12 @@
+'use strict';
+
+function escape(input) {
+  return String(input)
+        .replace(/&/g, '&amp;')
+        .replace(/</g, '&lt;')
+        .replace(/>/g, '&gt;')
+        .replace(/"/g, '&quot;')
+        .replace(/'/g, '&#x27;');
+}
+
+module.exports = escape;
